form-memory/backend/requirements.txt
fastapi
uvicorn
lxml
python-multipart
python-docx
python-dotenv
openai
